requirements.txt
gradio
transformers
peft
accelerate
bitsandbytes
sentencepiece
python-dotenv
langchain
